新增了SEFI项目的setup文件以用于发行；变更了文件夹名称；变更了requirements.txt，现使其拆分为SEFI和Lib两个，以规范安装行为；修复了部分错误引用问题；更新了排除文件；修复了部分不规范声明为Python软件包的问题

+++ canary_sefi/handler/image_handler/img_utils.py
@@ -11,7 +11,7 @@ def img_size_uniform_fix(ori_img, target_img, use_raw_nparray_data):
     if use_raw_nparray_data:
         img_type = np.float32
     else:
-        img_type = np.int
+        img_type = int
     ori_h, ori_w = ori_img.shape[0], ori_img.shape[1]
     adv_h, adv_w = target_img.shape[0], target_img.shape[1]
     if ori_h != adv_h or ori_w != adv_w:
